requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Faker==15.3.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
pandas==1.5.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
